MXS-2909: Add cluster accessors to Service
The initial assignment that is done at startup must not update the set of
targets the service uses as the addition of servers to a monitor will
trigger their addition to the service as well. All changes to the cluster
of a service done at runtime must update the set of used
servers. Currently the cluster is assigned only on startup.