Move active flag and lag related variables into Server
To keep things consistent, the ping variable is read without atomic
operations. This is safe as there is only one writer (one monitor per
server).

The deactivate method is exposed in the public class as clustrixmon needs
it. Instead, it should destroy the server if it is indeed permanently
disabled.