Make SERVER an abstract class
Moved the rest of the variables in to Server and made the rest of the
functions virtual.

The one-time flag used to detect TLS errors was removed to avoid
introducing it into Server. The flag isn't critical and removing it could
even be considered an improvement: If the backend TLS setup change it will
now be detected.